[WFCORE-5966] Rename WildflyTestRunner to WildFlyRunner
